Delete CborHelpers.netcoreapp.cs, merge into single CborHelpers.cs
Move all TFM-specific CborHelpers methods into one file with #if NET
guards. Delete the netcoreapp-only partial class file.
