--- src/libraries/System.Formats.Cbor/src/System/Formats/Cbor/CborHelpers.netcoreapp.cs
@@ -1,31 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using System.Buffers;
-using System.Buffers.Binary;
-using System.Numerics;
-using System.Runtime.CompilerServices;
-
-namespace System.Formats.Cbor
-{
-    internal static partial class CborHelpers
-    {
-        public static readonly DateTimeOffset UnixEpoch = DateTimeOffset.UnixEpoch;
-
-        public static BigInteger CreateBigIntegerFromUnsignedBigEndianBytes(byte[] bytes)
-            => new BigInteger(bytes, isUnsigned: true, isBigEndian: true);
-
-        public static byte[] CreateUnsignedBigEndianBytesFromBigInteger(BigInteger value)
-            => value.ToByteArray(isUnsigned: true, isBigEndian: true);
-
-        public static void GetBitsFromDecimal(decimal d, Span<int> destination)
-            => decimal.GetBits(d, destination);
-
-        public static string BuildStringFromIndefiniteLengthTextString<TState>(int length, TState state, SpanAction<char, TState> action)
-            => string.Create(length, state, action);
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Half ReadHalfBigEndian(ReadOnlySpan<byte> source)
-            => BinaryPrimitives.ReadHalfBigEndian(source);
-    }
-}
